made ui changes
